+++ packages/Chem/scripts/find_rgroups_with_core.py
@@ -17,6 +17,7 @@ mols = []
 r_group_map = dict()
 ps = rdRGroupDecomposition.RGroupDecompositionParameters()
 ps.onlyMatchAtRGroups = onlyMatchAtRGroups
+ps.matchingStrategy = rdRGroupDecomposition.RGroupMatching.Greedy
 core = Chem.MolFromMolBlock(core, sanitize = True) if ("M  END" in core) else Chem.MolFromSmarts(core)
 if core is not None:
     for n in range(0, length):